+++ workspace/solutions/kotlin/lib/src/test/kotlin/s1/s8/s5/s6/s_1856_test.kt
@@ -1,0 +1,16 @@
+package s1.s8.s5.s6
+
+import utils.T
+import kotlin.test.Test
+import s1.s8.s5.s6.s_1856_00.Solution as s_1856_00
+
+class s_1856_test {
+    @Test
+    fun test() {
+        val t = T(
+            this::class,
+            s_1856_00::class,
+        )
+        t.run()
+    }
+}