+++ workspace/solutions/rust/src/s1/s8/mod.rs
@@ -1,2 +1,3 @@
+pub mod s5;
 pub mod s7;
 pub mod s9;