+++ workspace/solutions/rust/src/s1/s8/s5/s6/mod.rs
@@ -1,0 +1,5 @@
+pub mod s_1856_00;
+pub mod s_1856_01;
+pub mod s_1856_test;
+
+pub type TSolution = fn(Vec<i32>) -> i32;